Refactored to remove mutationObserver toggling, made revisions based on provided comments, updated documentation
