+++ app/dom_event.html
@@ -34,10 +34,6 @@
                 cwr('enable');
             }
 
-            function allowDynamicDomEventListeners(allow) {
-                cwr('allowDynamicDomEventListeners', allow);
-            }
-
             function dynamicallyCreateButton() {
                 const parentButton = document.getElementById(
                     'dynamicallyCreateButton'
@@ -97,19 +93,6 @@
         <button id="dispatch" onclick="dispatch()">Dispatch</button>
         <button id="clearRequestResponse" onclick="clearRequestResponse()">
             Clear
-        </button>
-        <hr />
-        <button
-            id="toggleOnDynamicListeners"
-            onclick="allowDynamicDomEventListeners(true)"
-        >
-            Toggle On Dynamic Event Listeners
-        </button>
-        <button
-            id="toggleOffDynamicListeners"
-            onclick="allowDynamicDomEventListeners(false)"
-        >
-            Toggle Off Dynamic Event Listeners
         </button>
         <hr />
         <span id="request"></span>
