+++ docs/cdn_commands.md
@@ -53,7 +53,6 @@ if(awsCreds) awsRum.setAwsCredentials(credentialProvider);
 | Command | Parameter Type | Example <div style="width:265px"></div> | Description |
 | --- | --- | --- | --- |
 | allowCookies | Boolean | `cwr('allowCookies', true);`<br/><br/>`awsRum.allowCookies(true)` | Enable the web client to set and read two cookies: a session cookie named `cwr_s` and a user cookie named `cwr_u`.<br/><br/>`cwr_s` stores session data including an anonymous session ID (uuid v4) created by the web client. This allows CloudWatch RUM to compute sessionized metrics like errors per session.<br/><br/>`cwr_u` stores an anonymous user ID (uuid v4) created by the web client. This allows CloudWatch RUM to count return visitors.<br/><br/>`true`: the web client will use cookies<br/>`false`: the web client will not use cookies
-| allowDynamicDomEventListeners | Boolean | `cwr('allowDynamicDomEventListeners', true);`<br/><br/>`awsRum.allowDynamicDomEventListeners(true)` | Enable the web client to listen to dynamically added DOM elements by enabling [`MutationObserver`](https://developer.mozilla.org/en-US/docs/Web/API/MutationObserver). <br/><br/>When MutationObserver is enabled, event listeners for the target events will be attached to the DOM elements that are added to the DOM after the initial page load. <br/><br/>Enabling MutationObserver does not seem to have a significant impact on application performance, adding approximately 4ms of latency every 1000 target DOM elements. If the number of target DOM elements were to reach a significantly large number, it may have an impact on application performance. <br/><br/>`true`: the web client will enable MutationObserver<br/>`false`: the web client will not enable MutationObserver
 | disable | None | `cwr('disable');`<br/><br/>`awsRum.disable();` | Stop recording and dispatching RUM events.
 | dispatch | None | `cwr('dispatch');`<br/><br/>`awsRum.dispatch();` | Flush RUM events from the cache and dispatch them to CloudWatch RUM using [`fetch`](https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API). 
 | dispatchBeacon | None | `cwr('dispatchBeacon');`<br/><br/>`awsRum.dispatchBeacon();` | Flush RUM events from the cache and dispatch them to CloudWatch RUM using [`sendBeacon`](https://developer.mozilla.org/en-US/docs/Web/API/Beacon_API).
