+++ docs/configuration.md
@@ -21,7 +21,6 @@ For example, the config object may look similar to the following:
 | Field Name | Type | Default | Description |
 | --- | --- | --- | --- |
 | allowCookies | Boolean | `false` | Enable the web client to set and read two cookies: a session cookie named `cwr_s` and a user cookie named `cwr_u`.<br/><br/>`cwr_s` stores session data including an anonymous session ID (uuid v4) created by the web client. This allows CloudWatch RUM to compute sessionized metrics like errors per session.<br/><br/>`cwr_u` stores an anonymous user ID (uuid v4) created by the web client. This allows CloudWatch RUM to count return visitors.<br/><br/>`true`: the web client will use cookies<br/>`false`: the web client will not use cookies. |
-| allowDynamicDomEventListeners | Boolean | `false` | Enable the web client to listen to dynamically added DOM elements by enabling [`MutationObserver`](https://developer.mozilla.org/en-US/docs/Web/API/MutationObserver). <br/><br/>When MutationObserver is enabled, event listeners for the target events will be attached to the DOM elements that are added to the DOM after the initial page load. <br/><br/>Enabling MutationObserver does not seem to have a significant impact on application performance, adding approximately 4ms of latency every 1000 target DOM elements. If the number of target DOM elements were to reach a significantly large number, it may have an impact on application performance. <br/><br/>`true`: the web client will enable MutationObserver<br/>`false`: the web client will not enable MutationObserver |
 | cookieAttributes | [CookieAttributes](#cookieattributes) | `{ domain: window.location.hostname, path: '/', sameSite: 'Strict', secure: true } ` | Cookie attributes are applied to all cookies stored by the web client, including `cwr_s` and `cwr_u`. |
 | disableAutoPageView | Boolean | `false` | When this field is `false`, the web client will automatically record page views.<br/><br/>By default, the web client records page views when (1) the page first loads and (2) the browser's [history API](https://developer.mozilla.org/en-US/docs/Web/API/History_API) is called. The page ID is `window.location.pathname`.<br/><br/>In some cases, the web client's instrumentation will not record the desired page ID. In this case, the web client's page view automation must be disabled using the `disableAutoPageView` configuration, and the application must be instrumented to record page views using the `recordPageView` command. |
 | enableRumClient | Boolean | `true` | When this field is `true`, the web client will record and dispatch RUM events. |
@@ -92,6 +91,7 @@ telemetries: [
 
 | Name | Type | Default | Description |
 | --- | --- | --- | --- |
+| enableMutationObserver | Boolean | `false` | Enable the web client to listen to dynamically added DOM elements by enabling [`MutationObserver`](https://developer.mozilla.org/en-US/docs/Web/API/MutationObserver). <br/><br/>When MutationObserver is enabled, event listeners for the target events will be attached to the DOM elements that are added to the DOM after the initial page load. <br/><br/>Enabling MutationObserver does not seem to have a significant impact on application performance, adding approximately 4ms of latency every 1000 target DOM elements. If the number of target DOM elements were to reach a significantly large number, it may have an impact on application performance. <br/><br/>`true`: the web client will enable MutationObserver<br/>`false`: the web client will not enable MutationObserver |
 | events | Array | `[]` | An array of target DOM events to record. Each DOM event is defined by an *event* and a *selector*. The event must be a [DOM event](https://www.w3schools.com/jsref/dom_obj_event.asp). The selector must be one of (1) `cssLocator`, (2) `elementId` or (3) `element`.<br/><br/>When two or more selectors are provided for a target DOM event, only one selector will be used. The selectors will be honored with the following precedence: (1) `cssLocator`, (2) `elementId` or (3) `element`. For example, if both `cssLocator` and `elementId` are provided, only the `cssLocator` selector will be used.<br/><br/>**Examples:**<br/>Record all elements identified by CSS selector `[label="label1"]`:<br/> `[{ event: 'click', cssLocator: '[label="label1"]' }]`<br/><br/>Record a single element with ID `mybutton`:<br/>`[{ event: 'click', elementId: 'mybutton' }]`<br/><br/>Record a complete clickstream<br/>`[{ event: 'click', element: document }]`. |
 
 ## Performance
