+++ src/orchestration/Orchestration.ts
@@ -57,7 +57,6 @@ export type PartialCookieAttributes = {
 
 export type PartialConfig = {
     allowCookies?: boolean;
-    allowDynamicDomEventListeners?: boolean;
     batchLimit?: number;
     clientBuilder?: ClientBuilder;
     cookieAttributes?: PartialCookieAttributes;
@@ -106,7 +105,6 @@ export const defaultCookieAttributes = (): CookieAttributes => {
 export const defaultConfig = (cookieAttributes: CookieAttributes): Config => {
     return {
         allowCookies: false,
-        allowDynamicDomEventListeners: false,
         batchLimit: 100,
         cookieAttributes,
         disableAutoPageView: false,
@@ -138,7 +136,6 @@ export type CookieAttributes = {
 
 export type Config = {
     allowCookies: boolean;
-    allowDynamicDomEventListeners: boolean;
     batchLimit: number;
     clientBuilder?: ClientBuilder;
     cookieAttributes: CookieAttributes;
@@ -284,13 +281,6 @@ export class Orchestration {
      */
     public allowCookies(allow: boolean) {
         this.config.allowCookies = allow;
-    }
-
-    /**
-     * @param allow when {@code false}, the RUM web client will not listen to dynamically added DOM targets.
-     */
-    public allowDynamicDomEventListeners(allow: boolean) {
-        this.pluginManager.allowDynamicDomEventListeners(allow);
     }
 
     /**
