+++ webpack/webpack.dev.js
@@ -11,8 +11,8 @@ module.exports = merge(common, {
         loader_standard: './src/loader/loader-standard.js',
         loader_page_event: './src/loader/loader-page-event.js',
         loader_dom_event: './src/loader/loader-dom-event.js',
-        loader_dom_event_dynamic_handlers_enabled:
-            './src/loader/loader-dom-event-dynamic-handlers-enabled.js',
+        loader_dom_event_mutation_observer_enabled:
+            './src/loader/loader-dom-event-mutation-observer-enabled.js',
         loader_ingestion: './src/loader/loader-ingestion.js',
         loader_js_error_event: './src/loader/loader-js-error-event.js',
         loader_http_fetch_event: './src/loader/loader-http-fetch-event.js',
